Share v2 endpoint coverage logic
